+++ slurmweb/apps/agent.py
@@ -91,3 +91,7 @@ class SlurmwebAppAgent(SlurmwebWebApp, RFLTokenizedRBACWebApp):
         else:
             logger.warning("Caching is disabled")
             self.cache = None
+
+        # Default RacksDB infrastructure is the cluster name.
+        if self.settings.racksdb.infrastructure is None:
+            self.settings.racksdb.infrastructure = self.settings.service.cluster